Remove using BuildStatus for TaskRunStatus ✂️
Using directly Pod to build TaskRunStatus
