+++ pkg/reconciler/v1alpha1/taskrun/resources/pod_test.go
@@ -29,9 +29,7 @@ import (
 	fakek8s "k8s.io/client-go/kubernetes/fake"
 
 	v1alpha1 "github.com/knative/build/pkg/apis/build/v1alpha1"
-	"github.com/knative/build/pkg/system"
 	"github.com/knative/pkg/apis"
-	duckv1alpha1 "github.com/knative/pkg/apis/duck/v1alpha1"
 	"github.com/tektoncd/pipeline/test/names"
 )
 
@@ -306,288 +304,6 @@ func TestMakePod(t *testing.T) {
 			}
 			if d := cmp.Diff(got.Annotations, wantAnnotations); d != "" {
 				t.Errorf("Diff annotations:\n%s", d)
-			}
-		})
-	}
-}
-
-func TestBuildStatusFromPod(t *testing.T) {
-	for _, c := range []struct {
-		desc      string
-		podStatus corev1.PodStatus
-		buildSpec v1alpha1.BuildSpec
-		want      v1alpha1.BuildStatus
-	}{{
-		desc:      "empty",
-		podStatus: corev1.PodStatus{},
-		buildSpec: v1alpha1.BuildSpec{},
-		want:      v1alpha1.BuildStatus{},
-	}, {
-		desc: "ignore-creds-init",
-		podStatus: corev1.PodStatus{
-			InitContainerStatuses: []corev1.ContainerStatus{{
-				// creds-init; ignored
-			}},
-			ContainerStatuses: []corev1.ContainerStatus{{
-				Name: "state-name",
-				State: corev1.ContainerState{
-					Terminated: &corev1.ContainerStateTerminated{
-						ExitCode: 123,
-					},
-				},
-			}},
-		},
-		buildSpec: v1alpha1.BuildSpec{
-			// no sources.
-		},
-		want: v1alpha1.BuildStatus{
-			StepsCompleted: []string{"state-name"},
-			StepStates: []corev1.ContainerState{{
-				Terminated: &corev1.ContainerStateTerminated{
-					ExitCode: 123,
-				},
-			}},
-		},
-	}, {
-		desc: "ignore-creds-init-and-source",
-		podStatus: corev1.PodStatus{
-			InitContainerStatuses: []corev1.ContainerStatus{{
-				// creds-init; ignored.
-			}, {
-				// git-init; ignored.
-			}},
-			ContainerStatuses: []corev1.ContainerStatus{{
-				Name: "state-name",
-				State: corev1.ContainerState{
-					Terminated: &corev1.ContainerStateTerminated{
-						ExitCode: 123,
-					},
-				},
-			}},
-		},
-		buildSpec: v1alpha1.BuildSpec{
-			Source: &v1alpha1.SourceSpec{
-				Git: &v1alpha1.GitSourceSpec{
-					Url:      "example.com",
-					Revision: "master",
-				},
-			},
-		},
-		want: v1alpha1.BuildStatus{
-			StepsCompleted: []string{"state-name"},
-			StepStates: []corev1.ContainerState{{
-				Terminated: &corev1.ContainerStateTerminated{
-					ExitCode: 123,
-				},
-			}},
-		},
-	}, {
-		desc: "ignore-creds-init-and-multiple-sources",
-		podStatus: corev1.PodStatus{
-			InitContainerStatuses: []corev1.ContainerStatus{{
-				// creds-init; ignored.
-			}, {
-				// first git-init; ignored.
-			}, {
-				// second git-init; ignored.
-			}},
-			ContainerStatuses: []corev1.ContainerStatus{{
-				Name: "state-name",
-				State: corev1.ContainerState{
-					Terminated: &corev1.ContainerStateTerminated{
-						ExitCode: 123,
-					},
-				},
-			}},
-		},
-		buildSpec: v1alpha1.BuildSpec{
-			Sources: []v1alpha1.SourceSpec{{
-				Git: &v1alpha1.GitSourceSpec{
-					Url:      "example.com",
-					Revision: "master",
-				},
-			}, {
-				Git: &v1alpha1.GitSourceSpec{
-					Url:      "yahoo.com",
-					Revision: "so cool",
-				},
-			}},
-		},
-		want: v1alpha1.BuildStatus{
-			StepsCompleted: []string{"state-name"},
-			StepStates: []corev1.ContainerState{{
-				Terminated: &corev1.ContainerStateTerminated{
-					ExitCode: 123,
-				},
-			}},
-		},
-	}, {
-		desc:      "success",
-		podStatus: corev1.PodStatus{Phase: corev1.PodSucceeded},
-		want: v1alpha1.BuildStatus{
-			Conditions: []duckv1alpha1.Condition{{
-				Type:   v1alpha1.BuildSucceeded,
-				Status: corev1.ConditionTrue,
-			}},
-		},
-	}, {
-		desc:      "running",
-		podStatus: corev1.PodStatus{Phase: corev1.PodRunning},
-		want: v1alpha1.BuildStatus{
-			Conditions: []duckv1alpha1.Condition{{
-				Type:   v1alpha1.BuildSucceeded,
-				Status: corev1.ConditionUnknown,
-				Reason: "Building",
-			}},
-		},
-	}, {
-		desc: "failure-terminated",
-		podStatus: corev1.PodStatus{
-			Phase:                 corev1.PodFailed,
-			InitContainerStatuses: []corev1.ContainerStatus{{
-				// creds-init status; ignored
-			}},
-			ContainerStatuses: []corev1.ContainerStatus{{
-				Name:    "status-name",
-				ImageID: "image-id",
-				State: corev1.ContainerState{
-					Terminated: &corev1.ContainerStateTerminated{
-						ExitCode: 123,
-					},
-				},
-			}},
-		},
-		want: v1alpha1.BuildStatus{
-			StepsCompleted: []string{"status-name"},
-			StepStates: []corev1.ContainerState{{
-				Terminated: &corev1.ContainerStateTerminated{
-					ExitCode: 123,
-				},
-			}},
-			Conditions: []duckv1alpha1.Condition{{
-				Type:    v1alpha1.BuildSucceeded,
-				Status:  corev1.ConditionFalse,
-				Message: `build step "status-name" exited with code 123 (image: "image-id"); for logs run: kubectl -n knative-build logs pod -c status-name`,
-			}},
-		},
-	}, {
-		desc: "failure-message",
-		podStatus: corev1.PodStatus{
-			Phase:   corev1.PodFailed,
-			Message: "boom",
-		},
-		want: v1alpha1.BuildStatus{
-			Conditions: []duckv1alpha1.Condition{{
-				Type:    v1alpha1.BuildSucceeded,
-				Status:  corev1.ConditionFalse,
-				Message: "boom",
-			}},
-		},
-	}, {
-		desc:      "failure-unspecified",
-		podStatus: corev1.PodStatus{Phase: corev1.PodFailed},
-		want: v1alpha1.BuildStatus{
-			Conditions: []duckv1alpha1.Condition{{
-				Type:    v1alpha1.BuildSucceeded,
-				Status:  corev1.ConditionFalse,
-				Message: "build failed for unspecified reasons.",
-			}},
-		},
-	}, {
-		desc: "pending-waiting-message",
-		podStatus: corev1.PodStatus{
-			Phase:                 corev1.PodPending,
-			InitContainerStatuses: []corev1.ContainerStatus{{
-				// creds-init status; ignored
-			}},
-			ContainerStatuses: []corev1.ContainerStatus{{
-				Name: "status-name",
-				State: corev1.ContainerState{
-					Waiting: &corev1.ContainerStateWaiting{
-						Message: "i'm pending",
-					},
-				},
-			}},
-		},
-		want: v1alpha1.BuildStatus{
-			StepStates: []corev1.ContainerState{{
-				Waiting: &corev1.ContainerStateWaiting{
-					Message: "i'm pending",
-				},
-			}},
-			Conditions: []duckv1alpha1.Condition{{
-				Type:    v1alpha1.BuildSucceeded,
-				Status:  corev1.ConditionUnknown,
-				Reason:  "Pending",
-				Message: `build step "status-name" is pending with reason "i'm pending"`,
-			}},
-		},
-	}, {
-		desc: "pending-pod-condition",
-		podStatus: corev1.PodStatus{
-			Phase: corev1.PodPending,
-			Conditions: []corev1.PodCondition{{
-				Status:  corev1.ConditionUnknown,
-				Type:    "the type",
-				Message: "the message",
-			}},
-		},
-		want: v1alpha1.BuildStatus{
-			Conditions: []duckv1alpha1.Condition{{
-				Type:    v1alpha1.BuildSucceeded,
-				Status:  corev1.ConditionUnknown,
-				Reason:  "Pending",
-				Message: `pod status "the type":"Unknown"; message: "the message"`,
-			}},
-		},
-	}, {
-		desc: "pending-message",
-		podStatus: corev1.PodStatus{
-			Phase:   corev1.PodPending,
-			Message: "pod status message",
-		},
-		want: v1alpha1.BuildStatus{
-			Conditions: []duckv1alpha1.Condition{{
-				Type:    v1alpha1.BuildSucceeded,
-				Status:  corev1.ConditionUnknown,
-				Reason:  "Pending",
-				Message: "pod status message",
-			}},
-		},
-	}, {
-		desc:      "pending-no-message",
-		podStatus: corev1.PodStatus{Phase: corev1.PodPending},
-		want: v1alpha1.BuildStatus{
-			Conditions: []duckv1alpha1.Condition{{
-				Type:    v1alpha1.BuildSucceeded,
-				Status:  corev1.ConditionUnknown,
-				Reason:  "Pending",
-				Message: "Pending",
-			}},
-		},
-	}} {
-		t.Run(c.desc, func(t *testing.T) {
-			now := metav1.Now()
-			p := &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:              "pod",
-					Namespace:         system.Namespace,
-					CreationTimestamp: now,
-				},
-				Status: c.podStatus,
-			}
-			got := BuildStatusFromPod(p, c.buildSpec)
-
-			// Common traits, set for test case brevity.
-			c.want.Cluster = &v1alpha1.ClusterSpec{
-				PodName:   "pod",
-				Namespace: system.Namespace,
-			}
-			c.want.Builder = v1alpha1.ClusterBuildProvider
-			c.want.StartTime = &now
-
-			if d := cmp.Diff(got, c.want, ignoreVolatileTime); d != "" {
-				t.Errorf("Diff:\n%s", d)
 			}
 		})
 	}
