+++ pkg/reconciler/v1alpha1/taskrun/taskrun_test.go
@@ -847,164 +847,6 @@ func TestReconcileBuildUpdateStatus(t *testing.T) {
 	}
 }
 
-func TestUpdateStatusFromBuildStatus(t *testing.T) {
-	completed := corev1.ContainerState{
-		Terminated: &corev1.ContainerStateTerminated{ExitCode: 0, Reason: "success"},
-	}
-	waiting := corev1.ContainerState{
-		Waiting: &corev1.ContainerStateWaiting{
-			Message: "foo",
-			Reason:  "bar",
-		},
-	}
-	failed := corev1.ContainerState{
-		Terminated: &corev1.ContainerStateTerminated{ExitCode: 127, Reason: "oh-my-lord"},
-	}
-	startTime := metav1.NewTime(time.Date(2018, time.November, 10, 23, 0, 0, 0, time.UTC))
-	completionTime := metav1.NewTime(time.Date(2018, time.November, 10, 23, 8, 0, 0, time.UTC))
-	for _, tc := range []struct {
-		name           string
-		buildStatus    buildv1alpha1.BuildStatus
-		expectedStatus v1alpha1.TaskRunStatus
-	}{{
-		name:        "empty build status",
-		buildStatus: buildv1alpha1.BuildStatus{},
-		expectedStatus: v1alpha1.TaskRunStatus{
-			Conditions: []duckv1alpha1.Condition{{
-				Type:    duckv1alpha1.ConditionSucceeded,
-				Status:  corev1.ConditionUnknown,
-				Reason:  reasonRunning,
-				Message: reasonRunning,
-			}},
-			Steps: []v1alpha1.StepState{},
-		},
-	}, {
-		name: "build validate failed",
-		buildStatus: buildv1alpha1.BuildStatus{
-			Conditions: []duckv1alpha1.Condition{{
-				Type:    duckv1alpha1.ConditionSucceeded,
-				Status:  corev1.ConditionFalse,
-				Reason:  "BuildValidationFailed",
-				Message: `serviceaccounts "missing-sa" not-found`,
-			}},
-		},
-		expectedStatus: v1alpha1.TaskRunStatus{
-			Conditions: []duckv1alpha1.Condition{{
-				Type:    duckv1alpha1.ConditionSucceeded,
-				Status:  corev1.ConditionFalse,
-				Reason:  "BuildValidationFailed",
-				Message: `serviceaccounts "missing-sa" not-found`,
-			}},
-			Steps: []v1alpha1.StepState{},
-		},
-	}, {
-		name: "running build status",
-		buildStatus: buildv1alpha1.BuildStatus{
-			StartTime: &startTime,
-			StepStates: []corev1.ContainerState{
-				waiting,
-			},
-			Conditions: []duckv1alpha1.Condition{{
-				Type:    duckv1alpha1.ConditionSucceeded,
-				Reason:  "Running build",
-				Message: "Running build",
-			}},
-			Cluster: &buildv1alpha1.ClusterSpec{
-				Namespace: "default",
-				PodName:   "im-am-the-pod",
-			},
-		},
-		expectedStatus: v1alpha1.TaskRunStatus{
-			Conditions: []duckv1alpha1.Condition{{
-				Type:    duckv1alpha1.ConditionSucceeded,
-				Reason:  "Running build",
-				Message: "Running build",
-			}},
-			Steps: []v1alpha1.StepState{
-				{ContainerState: *waiting.DeepCopy()},
-			},
-			PodName:   "im-am-the-pod",
-			StartTime: &startTime,
-		},
-	}, {
-		name: "completed build status (success)",
-		buildStatus: buildv1alpha1.BuildStatus{
-			StartTime:      &startTime,
-			CompletionTime: &completionTime,
-			StepStates: []corev1.ContainerState{
-				completed,
-			},
-			Conditions: []duckv1alpha1.Condition{{
-				Type:    duckv1alpha1.ConditionSucceeded,
-				Status:  corev1.ConditionTrue,
-				Reason:  "Build succeeded",
-				Message: "Build succeeded",
-			}},
-			Cluster: &buildv1alpha1.ClusterSpec{
-				Namespace: "default",
-				PodName:   "im-am-the-pod",
-			},
-		},
-		expectedStatus: v1alpha1.TaskRunStatus{
-			Conditions: []duckv1alpha1.Condition{{
-				Type:    duckv1alpha1.ConditionSucceeded,
-				Status:  corev1.ConditionTrue,
-				Reason:  "Build succeeded",
-				Message: "Build succeeded",
-			}},
-			Steps: []v1alpha1.StepState{
-				{ContainerState: *completed.DeepCopy()},
-			},
-			PodName:        "im-am-the-pod",
-			StartTime:      &startTime,
-			CompletionTime: &completionTime,
-		},
-	}, {
-		name: "completed build status (failure)",
-		buildStatus: buildv1alpha1.BuildStatus{
-			StartTime:      &startTime,
-			CompletionTime: &completionTime,
-			StepStates: []corev1.ContainerState{
-				completed,
-				failed,
-			},
-			Conditions: []duckv1alpha1.Condition{{
-				Type:    duckv1alpha1.ConditionSucceeded,
-				Status:  corev1.ConditionFalse,
-				Reason:  "Build failed",
-				Message: "Build failed",
-			}},
-			Cluster: &buildv1alpha1.ClusterSpec{
-				Namespace: "default",
-				PodName:   "im-am-the-pod",
-			},
-		},
-		expectedStatus: v1alpha1.TaskRunStatus{
-			Conditions: []duckv1alpha1.Condition{{
-				Type:    duckv1alpha1.ConditionSucceeded,
-				Status:  corev1.ConditionFalse,
-				Reason:  "Build failed",
-				Message: "Build failed",
-			}},
-			Steps: []v1alpha1.StepState{
-				{ContainerState: *completed.DeepCopy()},
-				{ContainerState: *failed.DeepCopy()},
-			},
-			PodName:        "im-am-the-pod",
-			StartTime:      &startTime,
-			CompletionTime: &completionTime,
-		},
-	}} {
-		t.Run(tc.name, func(t *testing.T) {
-			taskRun := &v1alpha1.TaskRun{}
-			updateStatusFromBuildStatus(taskRun, tc.buildStatus)
-			if d := cmp.Diff(taskRun.Status, tc.expectedStatus, ignoreLastTransitionTime); d != "" {
-				t.Errorf("-want, +got: %v", d)
-			}
-		})
-	}
-}
-
 func TestCreateRedirectedBuild(t *testing.T) {
 
 	tr := tb.TaskRun("tr", "tr", tb.TaskRunSpec(
